+++ mvc/src/main/java/com/interface21/web/DispatcherServletInitializer.java
@@ -1,9 +1,8 @@
-package com.techcourse;
+package com.interface21.web;
 
 import jakarta.servlet.ServletContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import com.interface21.web.WebApplicationInitializer;
 
 /**
  * Base class for {@link WebApplicationInitializer}
