+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/HandlerExecution.java
@@ -1,5 +1,6 @@
 package com.interface21.webmvc.servlet.mvc.tobe;
 
+import com.interface21.webmvc.servlet.ModelAndView;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.slf4j.Logger;
@@ -21,9 +22,9 @@ public class HandlerExecution {
         log.info("Controller Instance : {}, Controller Method : {}",controller,method);
     }
 
-    public Object handle(final HttpServletRequest request, final HttpServletResponse response) {
+    public ModelAndView handle(final HttpServletRequest request, final HttpServletResponse response) {
         try {
-            return method.invoke(controller, request, response);
+            return (ModelAndView) method.invoke(controller, request, response);
         } catch (final IllegalAccessException | InvocationTargetException e) {
             throw new FailHandlerExecuteException(method, e);
         }
